requirements.txt
google-genai>=0.2.0
google-cloud-aiplatform>=1.40.0
google-cloud-firestore>=2.13.0
python-dotenv>=1.0.0
flask>=3.0.0
pandas>=2.1.0
